+++ rules-tests/DeadCode/Rector/Stmt/RemoveConditionExactReturnRector/Fixture/skip_different_return.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\Stmt\RemoveEmptyArrayConditionReturnRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\Stmt\RemoveConditionExactReturnRector\Fixture;
 
 final class SkipDifferentReturn
 {
