+++ rules-tests/DeadCode/Rector/Stmt/RemoveConditionExactReturnRector/config/configured_rule.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Rector\DeadCode\Rector\Stmt\RemoveConditionExactReturnRector;
+
+return RectorConfig::configure()
+    ->withRules([RemoveConditionExactReturnRector::class]);
