--- rules-tests/DeadCode/Rector/Stmt/RemoveEmptyArrayConditionReturnRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\DeadCode\Rector\Stmt\RemoveEmptyArrayConditionReturnRector;
-
-return RectorConfig::configure()
-    ->withRules([RemoveEmptyArrayConditionReturnRector::class]);
